remove task from dict when completed

+++ mygen.py
@@ -5,7 +5,7 @@ def load_model():
     model_id = "stabilityai/stable-diffusion-2-1"
     
     # Use the DPMSolverMultistepScheduler (DPM-Solver++) scheduler here instead
-    pipe = StableDiffusionPipeline.from_pretrained(model_id, torch_dtype=torch.float16)
+    pipe = StableDiffusionPipeline.from_pretrained(model_id, torch_dtype=torch.float32)
     pipe.scheduler = DPMSolverMultistepScheduler.from_config(pipe.scheduler.config)
     pipe = pipe.to("cuda")
     return pipe
